Consolidate configure controller
