+++ app/controllers/imports_controller.rb
@@ -1,3 +1,5 @@
+require "ostruct"
+
 class ImportsController < ApplicationController
   before_action :set_import, except: %i[ index new create ]
 
@@ -36,10 +38,22 @@ class ImportsController < ApplicationController
 
   def update_csv
     if @import.update(import_params)
-      redirect_to import_configure_path(@import), notice: "Import uploaded"
+      redirect_to configure_import_path(@import), notice: "Import uploaded"
     else
       flash.now[:error] = @import.errors.full_messages.to_sentence
       render :load, status: :unprocessable_entity
+    end
+  end
+
+  def configure
+  end
+
+  def update_mappings
+    if @import.update(import_params)
+      redirect_to import_clean_path(@import), notice: "Mappings saved"
+    else
+      flash.now[:error] = @import.errors.full_messages.first
+      render :show, status: :unprocessable_entity
     end
   end
 
@@ -50,6 +64,6 @@ class ImportsController < ApplicationController
     end
 
     def import_params
-      params.require(:import).permit(:raw_csv)
+      params.require(:import).permit(:raw_csv, column_mappings: [ :date, :merchant, :category, :amount ])
     end
 end
