--- app/helpers/imports/configures_helper.rb
@@ -1,2 +1,0 @@
-module Imports::ConfiguresHelper
-end
