--- app/helpers/imports/loads_helper.rb
@@ -1,2 +1,0 @@
-module Imports::LoadsHelper
-end
