--- app/helpers/imports/steps_helper.rb
@@ -1,2 +1,0 @@
-module Imports::StepsHelper
-end
