+++ app/views/imports/_nav.html.erb
@@ -4,7 +4,7 @@
     <div class="h-px bg-alpha-black-200 w-12"></div>
     <%= render partial: "imports/nav_item", locals: { name: "Import", path: import.persisted? ? load_import_path(import) : nil, index: 1, is_complete: false, is_current: false } %>
     <div class="h-px bg-alpha-black-200 w-12"></div>
-    <%= render partial: "imports/nav_item", locals: { name: "Setup", path: import.persisted? ? import_configure_path(import) : nil, index: 2, is_complete: false, is_current: false } %>
+    <%= render partial: "imports/nav_item", locals: { name: "Setup", path: import.persisted? ? configure_import_path(import) : nil, index: 2, is_complete: false, is_current: false } %>
     <div class="h-px bg-alpha-black-200 w-12"></div>
     <%= render partial: "imports/nav_item", locals: { name: "Clean", path: import.persisted? ? import_clean_path(import) : nil, index: 3, is_complete: false, is_current: false } %>
     <div class="h-px bg-alpha-black-200 w-12"></div>
