+++ app/views/imports/configure.html.erb
@@ -7,7 +7,7 @@
 <div>
   <h1 class="sr-only">Configure import</h1>
 
-  <%= form_with model: @import, url: import_configure_path(@import), method: :patch do |form| %>
+  <%= form_with model: @import, url: configure_import_path(@import) do |form| %>
     <div class="mb-4">
       <%= form.fields_for :column_mappings, OpenStruct.new(@import.column_mappings || @import.default_column_mappings) do |mappings| %>
         <%= mappings.text_field :date, label: "Date" %>
